Refactor the DownloadDropdown component to have dedicated container for state
